fix:details api successfully working

+++ ecommerceapp/serializer.py
@@ -57,48 +57,36 @@ class ImageSerializer(serializers.ModelSerializer):
 
 class DetailsSerializer(serializers.ModelSerializer):
     images=ImageSerializer(many=True, read_only=True)
+    unique_url = serializers.PrimaryKeyRelatedField(read_only=True)
+
     class Meta:
         model=Details
         fields=['unique_url','title','description','images']
 
     def create(self, validated_data):
+            # Extract title and description from validated data
+        title = validated_data.get('title', "")
+        description = validated_data.get('description', "")
+        details_instance = Details.objects.create(
+            unique_url=validated_data['unique_url'],
+            title=title,
+            description=description
+        )
         image_data = self.context['request'].FILES.getlist('file')
-        details_instance, created = Details.objects.update_or_create(**validated_data)
-
-        for image in image_data:    
+        for image in image_data:
             Images.objects.create(detail=details_instance, file=image)
-
         return details_instance
-    
     def update(self, instance, validated_data):
-        image_data = self.context['request'].FILES.getlist('file')
         instance.title = validated_data.get('title', instance.title)
         instance.description = validated_data.get('description', instance.description)
-
-        Images.objects.filter(detail=instance).delete()
-        
-        for image in image_data:
-            Images.objects.create(detail=instance, file=image)
+        instance.save()
+        image_data = self.context['request'].FILES.getlist('file')
+        if image_data:
+            Images.objects.filter(detail=instance).delete()
+            for image in image_data:
+                Images.objects.create(detail=instance, file=image)
 
         return instance
-
-    # def create(self, validated_data):
-    #     request = self.context['request']
-    #     images_data = request.FILES.getlist('images')
-    #     print(images_data,"helo")
-    #     details_instance, created = Details.objects.update_or_create(
-    #         unique_url=validated_data.pop('unique_url'),  # Find by unique_url
-    #         defaults=validated_data  # Update other fields
-        
-    #     )
-    #     # Handle image uploads
-    #     for image in images_data:
-    #         Images.objects.create(detail=details_instance, file=image)
-            
-    #     return details_instance
-
-
-
